Add support for ChaCha20Poly1305 RFC.
This adds support for the ChaCha20Poly1305 AEAD construction specified
in RFC7539.

It adds NewRFC and NewDraft. NewDraft behavoir is identical to the
previous New, and NewRFC is an implementation of RFC7539.

NewDraft returns an AEAD cipher that takes a 64 bit nonce. NewRFC
returns an AEAD cipher that takes a 96 bit nonce.

New is deprected and behaves identical to NewDraft.

Performance is identical.